backend/requirements.txt
flask==2.0.1
werkzeug==2.0.3
flask-cors==3.0.10
python-dotenv==0.19.0
google-generativeai==0.3.0
requests==2.26.0
gunicorn==20.1.0 